backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pymongo==4.6.1
requests==2.31.0
Pillow>=10.0.0  # Using latest version since we handle both textsize and textlength methods
psycopg2-binary==2.9.9  # PostgreSQL adapter 